+++ crates/winmd/src/file.rs
@@ -601,9 +601,9 @@ impl File {
         &self.tables[TableIndex::TypeDef as usize]
     }
 
-    pub(crate) fn nested_class_table(&self) -> &TableData {
-        &self.tables[TableIndex::NestedClass as usize]
-    }
+    // pub(crate) fn nested_class_table(&self) -> &TableData {
+    //     &self.tables[TableIndex::NestedClass as usize]
+    // }
 }
 
 fn section_from_rva(sections: &[ImageSectionHeader], rva: u32) -> &ImageSectionHeader {
