+++ crates/winmd/src/lib.rs
@@ -1,12 +1,54 @@
 //! A Windows Metadata (winmd) parser
 mod file;
-pub mod parsed;
 mod traits;
 mod type_reader;
 mod workspace;
+mod assembly_ref;
+mod attribute;
+mod blob;
+mod codes;
+mod constant;
+mod constant_value;
+mod element_type;
+mod field;
+mod flags;
+mod generic_param;
+mod impl_map;
+mod interface_impl;
+mod member_ref;
+mod method_def;
+mod module;
+mod module_ref;
+mod nested_class;
+mod param;
+mod row;
+mod type_def;
+mod type_ref;
+mod type_spec;
 
-pub use file::{File, TableIndex};
-pub use parsed::*;
+pub use assembly_ref::*;
+pub use attribute::*;
+pub use blob::*;
+pub use codes::*;
+pub use constant::*;
+pub use constant_value::*;
+pub use element_type::*;
+pub use field::*;
+pub use flags::*;
+pub use generic_param::*;
+pub use impl_map::*;
+pub use interface_impl::*;
+pub use member_ref::*;
+pub use method_def::*;
+pub use module::*;
+pub use module_ref::*;
+pub use nested_class::*;
+pub use param::*;
+pub use row::*;
+pub use type_def::*;
+pub use type_ref::*;
+pub use type_spec::*;
+pub use file::*;
 pub use traits::*;
 pub use type_reader::*;
 pub use workspace::*;
