--- crates/winmd/src/parsed/constant_value.rs
@@ -1,14 +1,0 @@
-#[derive(Debug)]
-pub enum ConstantValue {
-    U8(u8),
-    I8(i8),
-    U16(u16),
-    I16(i16),
-    U32(u32),
-    I32(i32),
-    U64(u64),
-    I64(i64),
-    F32(f32),
-    F64(f64),
-    String(String),
-}
