+++ crates/winmd/src/type_reader.rs
@@ -1,4 +1,4 @@
-use crate::*;
+use super::*;
 use std::cmp::Ordering;
 use std::collections::BTreeMap;
 use std::path::PathBuf;
