+++ crates/winmd/tests/stringable.rs
@@ -7,7 +7,7 @@ fn stringable() {
     let def = reader.expect_type_def(("Windows.Foundation", "IStringable"));
     assert!(def.name() == ("Windows.Foundation", "IStringable"));
 
-    let methods: Vec<winmd::parsed::MethodDef> = def.methods().collect();
+    let methods: Vec<winmd::MethodDef> = def.methods().collect();
     assert!(methods.len() == 1);
 
     let method = methods[0];
